PythonProject5/requirements.txt
numpy==1.26.4
soundfile==0.12.1
vosk==0.3.45
openai-whisper==20231117
python-Levenshtein==0.25.0
openpyxl==3.1.2
scipy==1.13.0
tqdm==4.66.2
transformers==4.41.2
# PyTorch 会自动安装兼容版本